+++ tenmo-server/src/main/java/com/techelevator/tenmo/dao/TransferStatusDao.java
@@ -6,6 +6,6 @@ public interface TransferStatusDao {
 
     TransferStatus getTransferStatusByDesc(String description);
 
-    TransferStatus getTransferStatusById(int transferStatusId);
+    TransferStatus getTransferStatusById(long transferStatusId);
 
 }
